Move utils so pipeline does not depend on core
