+++ rastervision2/examples/sample_pipeline2/sample_pipeline2.py
@@ -5,7 +5,7 @@ from rastervision2.pipeline.pipeline import Pipeline
 from rastervision2.pipeline.file_system import str_to_file, file_to_str
 from rastervision2.pipeline.pipeline_config import PipelineConfig
 from rastervision2.pipeline.config import register_config, Config
-from rastervision2.core.utils.misc import split_into_groups
+from rastervision2.pipeline.utils import split_into_groups
 
 
 @register_config('message_maker')
